refactor data retrieval into data service
